Update README.md
We Dedicate the News Portal "Bangladeshi Television's Network" is as an Ideal News Media Station for all the Mainstream Journalists of the World Including my Native Bangladesh. We Strongly Believe that we will be Able to rid the World of False Journalism or Yellow Journalism. Just as the Honorable Prime Minister of Our Bangladesh is Fighting to build a World free from the Effects of Climate Pollution, Our Fight is to Eradicate bad Journalism or Yellow Journalism from the World and Promote the Development and Progress of Mainstream Journalism.